creación del main, instacia de administrador

+++ src/main/java/org/example/herencia/Camioneta.java
@@ -3,14 +3,14 @@ package org.example.herencia;
 import org.example.modelos.Vehiculo;
 
 public class Camioneta extends Vehiculo {
-    private int capacidadCarga;
+    private double capacidadCarga;
 
-    public Camioneta(String idVehiculo, String marca, String modelo, int anno, double costoDiario, int capacidadCarga) {
+    public Camioneta(String idVehiculo, String marca, String modelo, int anno, double costoDiario, double capacidadCarga) {
         super(idVehiculo, marca, modelo, anno, costoDiario);
         this.capacidadCarga = capacidadCarga;
     }
 
-    public int getCapacidadCarga() {
+    public double getCapacidadCarga() {
         return capacidadCarga;
     }
 }
